feat: added screening generation

+++ back/importer/src/main/java/cinema/importer/models/ImportedMovie.java
@@ -15,5 +15,5 @@ public class ImportedMovie {
 
   String crew;
   double imDbRating;
-  double imDbRatingCount;
+  int imDbRatingCount;
 }
